[MODEL] Added proper support for the new "multi_fields" in the mapping DSL
This branch updates the `mapping` DSL to have nested `indexes` calls produce
embedded fields (instead of properties) for the core types where this is
appropriate.

This is allowed starting with [elasticsearch 1.0][es10mf] as a replacement for
`multi_field`.

[es10mf]: http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/_multi_fields.html

`type: "multi_field"` will still work as before; this change should not break
backwards compatibility.

Closes #138