[MODEL] [STORE] Fixed the problem where `document_type` configuration was not propagated to mapping
This PR tries to fix the bug described in #270

The problem was the following:

* Once a model extends Persistence::Model it sets up two attributes. As a consequence, the attribute macro instantiates a Elasticsearch::Model::Indexing::Mappings in the model. This instance contains an attribute @type that defaults to the model name in lowercase.

* Any other definition of document_type will override the document_type in the gateway delegate, but the mappings instance originally created in the previous step remains untouched, and hence in an inconsistent state.

What I did here is: Instead of simply forward document_type to the gateway, also change the mappings' document_type accordingly, leaving it in a correct state.

An integration test was modified, in order to have a regression of the fix.

Unit tests were added.

Fixes #270
Closes #366